Make executable paths platform agnostic
While the devkitPro path is easily accessible from the top of devkit_tools.py, the actual executable paths are farther down and can surprise users not on Windows with a path error when compiling.  The extension of the executable is now determined by a self.extension variable that does a platform.system() check. This check currently assumes that all other platforms besides Windows lack an extension for executable files.